backend/services/calendar_service.py
from google.oauth2.credentials import Credentials
from googleapiclient.discovery import build
from typing import Dict

class CalendarService:
    def __init__(self, client_id: str, client_secret: str):
        self.client_id = client_id
        self.client_secret = client_secret
    
    async def create_event(self, access_token: str, event_data: Dict) -> Dict:
        """Create Google Calendar event"""
        try:
            credentials = Credentials(token=access_token)
            service = build('calendar', 'v3', credentials=credentials)
            
            event = {
                'summary': event_data['title'],
                'description': event_data.get('description', ''),
                'start': {
                    'dateTime': event_data['start_time'],
                    'timeZone': 'America/New_York',
                },
                'end': {
                    'dateTime': event_data['end_time'],
                    'timeZone': 'America/New_York',
                },
                'reminders': {
                    'useDefault': False,
                    'overrides': [
                        {'method': 'popup', 'minutes': 30},
                    ],
                },
            }
            
            created_event = service.events().insert(
                calendarId='primary',
                body=event
            ).execute()
            
            return {
                'id': created_event['id'],
                'link': created_event.get('htmlLink'),
                'status': 'created'
            }
        except Exception as e:
            raise Exception(f"Calendar event creation failed: {str(e)}")
    
    def get_auth_url(self, redirect_uri: str) -> str:
        """Generate Google OAuth URL"""
        from google_auth_oauthlib.flow import Flow
        
        flow = Flow.from_client_config(
            {
                "web": {
                    "client_id": self.client_id,
                    "client_secret": self.client_secret,
                    "auth_uri": "https://accounts.google.com/o/oauth2/auth",
                    "token_uri": "https://oauth2.googleapis.com/token",
                }
            },
            scopes=['https://www.googleapis.com/auth/calendar.events'],
            redirect_uri=redirect_uri
        )
        
        auth_url, _ = flow.authorization_url(prompt='consent')
        return auth_url

