backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.0
anthropic==0.7.0
websockets==12.0
python-multipart==0.0.6
httpx==0.25.1
pydantic==2.5.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dateutil==2.8.2
aiofiles==23.2.1
google-genai==0.2.2

